북마크 추가 삭제까지 성공

+++ src/main/java/com/bitcamp/team_project_eco/bookmark/BookmarkRepository.java
@@ -1,7 +1,8 @@
 package com.bitcamp.team_project_eco.bookmark;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark,Integer>, CustomedBookmarkRepository {
+public interface BookmarkRepository extends JpaRepository<Bookmark,Long>, CustomedBookmarkRepository {
 
 }
