+++ src/main/java/com/bitcamp/team_project_eco/chargingStation/ChargingStationRepository.java
@@ -2,5 +2,5 @@ package com.bitcamp.team_project_eco.chargingStation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChargingStationRepository extends JpaRepository<ChargingStation,Integer>, CustomedChargingStationRepository {
+public interface ChargingStationRepository extends JpaRepository<ChargingStation,Long>, CustomedChargingStationRepository {
 }
