+++ src/main/java/com/bitcamp/team_project_eco/image/ImageRepository.java
@@ -2,5 +2,5 @@ package com.bitcamp.team_project_eco.image;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image,Integer>, CustomedImageRepository {
+public interface ImageRepository extends JpaRepository<Image,Long>, CustomedImageRepository {
 }
