+++ src/main/java/com/bitcamp/team_project_eco/purchase/PurchaseRepository.java
@@ -2,5 +2,5 @@ package com.bitcamp.team_project_eco.purchase;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRepository extends JpaRepository<Purchase,Integer>, CustomedPurchaseRepository {
+public interface PurchaseRepository extends JpaRepository<Purchase,Long>, CustomedPurchaseRepository {
 }
