+++ src/main/java/com/bitcamp/team_project_eco/wishlist/WishlistRepository.java
@@ -2,5 +2,5 @@ package com.bitcamp.team_project_eco.wishlist;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WishlistRepository extends JpaRepository<Wishlist,Integer>, CustomedWishlistRepository {
+public interface WishlistRepository extends JpaRepository<Wishlist,Long>, CustomedWishlistRepository {
 }
